survey/requirements.txt
celery==5.4.0
Django==4.2
django-environ==0.11.2
django-restframework==0.0.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
redis==5.2.0
django-cors-headers
djangorestframework-simplejwt
Pillow==11.0.0
